+++ src/main/java/edu/codespring/sportgh/dto/ProductOutDTO.java
@@ -8,7 +8,7 @@ public class ProductOutDTO {
     private Long id;
     private String name;
     private String description;
-    private Integer rentPrice;
+    private Double rentPrice;
     private Long subCategoryId;
     private Long userId;
 }
